[PIPE-10] renaming local app env module from lcl.py to local.py. cleanup docs and undo settings edit
